Apology commit
im gonna be soooo real with you guys, i totally thought facetious was the adjective used to describe someone that has a facade. like a fake face. this is so sad. i know now that it actually means something completely different. i think we should continute to be facetious on the internet. there should also be an adjective for facade.